Move node sources API impl out
