+++ src/api/implementation/cluster-explorer/v1.ts
@@ -2,7 +2,8 @@ import * as vscode from 'vscode';
 
 import { ClusterExplorerV1 } from "../../contract/cluster-explorer/v1";
 import { ExplorerExtender, ExplorerUICustomizer } from "../../../components/clusterexplorer/explorer.extension";
-import { KUBERNETES_EXPLORER_NODE_CATEGORY, ResourceFolder, ResourceNode, KubernetesExplorer, CustomResourceFolderNodeSource, CustomGroupingFolderNodeSource, NodeSourceImpl } from "../../../components/clusterexplorer/explorer";
+import { KUBERNETES_EXPLORER_NODE_CATEGORY, ResourceFolder, ResourceNode, KubernetesExplorer } from "../../../components/clusterexplorer/explorer";
+import { CustomResourceFolderNodeSource, CustomGroupingFolderNodeSource, NodeSourceImpl } from "../../../components/clusterexplorer/extension.nodesources";
 import { ClusterExplorerNode } from "../../../components/clusterexplorer/node";
 import { Kubectl } from "../../../kubectl";
 import { Host } from "../../../host";
